requirements.txt
# Contract Tool — Python dependencies
# Install: pip install -r requirements.txt
# Optional: copy .env.example to .env and set secrets (see docs/DEPLOYMENT.md)

streamlit>=1.28.0,<2
pandas>=2.0.0,<3
mysql-connector-python>=8.0.0,<9
openpyxl>=3.1.0,<4
bcrypt>=4.0.0,<5
python-dotenv>=1.0.0,<2
itsdangerous>=2.0.0,<3
extra-streamlit-components>=0.1.71,<0.2
